+++ R/distinct.R
@@ -82,7 +82,7 @@ distinct_prepare <- function(.data, vars, group_vars = character(), .keep_all = 
   }
 
   # If any calls, use mutate to add new columns, then distinct on those
-  computed_columns <- add_computed_columns(.data, vars)
+  computed_columns <- add_computed_columns(.data, vars, "distinct")
   .data <- computed_columns$data
   distinct_vars <- computed_columns$added_names