+++ R/group_data.R
@@ -5,11 +5,12 @@
 #'   The columns give the values of the grouping variables. The last column,
 #'   always called `.rows`, is a list of integer vectors that gives the
 #'   location of the rows in each group. You can retrieve just the grouping
-#'   data with `group_data()`, and just the locations with `group_rows()`.
+#'   data with `group_keys()`, and just the locations with `group_rows()`.
 #'
 #' * `group_indices()` returns an integer vector the same length as `.data`
 #'   that gives the group that each row belongs to (cf. `group_rows()` which
-#'   returns the rows which each group contains).
+#'   returns the rows which each group contains). `group_indices()` with no
+#'   argument is deprecated, superseded by [cur_group_id()].
 #'
 #' * `group_vars()` gives names of grouping variables as character vector;
 #'   `groups()` gives the names as a list of symbols.
@@ -29,11 +30,13 @@
 #' group_vars(df)
 #' group_rows(df)
 #' group_data(df)
+#' group_indices(df)
 #'
 #' gf <- group_by(df, x)
 #' group_vars(gf)
 #' group_rows(gf)
 #' group_data(gf)
+#' group_indices(gf)
 #' @export
 group_data <- function(.data) {
   UseMethod("group_data")