+++ src/filter.cpp
@@ -35,8 +35,9 @@ void stop_filter_incompatible_type(R_xlen_t i, SEXP quos, SEXP column_name, SEXP
 bool all_lgl_columns(SEXP data) {
   R_xlen_t nc = XLENGTH(data);
 
+  const SEXP* p_data = VECTOR_PTR_RO(data);
   for (R_xlen_t i = 0; i < nc; i++) {
-    if (TYPEOF(VECTOR_ELT(data, i)) != LGLSXP) return false;
+    if (TYPEOF(p_data[i]) != LGLSXP) return false;
   }
 
   return true;
@@ -73,12 +74,12 @@ void filter_check_type(SEXP res, R_xlen_t i, SEXP quos) {
     R_xlen_t ncol = XLENGTH(res);
     if (ncol == 0) return;
 
+    const SEXP* p_res = VECTOR_PTR_RO(res);
     for (R_xlen_t j=0; j<ncol; j++) {
-      SEXP res_j = VECTOR_ELT(res, j);
+      SEXP res_j = p_res[j];
       if (TYPEOF(res_j) != LGLSXP) {
         SEXP colnames = PROTECT(Rf_getAttrib(res, R_NamesSymbol));
-        SEXP colnames_j = PROTECT(Rf_allocVector(STRSXP, 1));
-        SET_STRING_ELT(colnames_j, 0, STRING_ELT(colnames, j));
+        SEXP colnames_j = PROTECT(Rf_ScalarString(STRING_ELT(colnames, j)));
         dplyr::stop_filter_incompatible_type(i, quos, colnames_j, res_j);
         UNPROTECT(2);
       }
@@ -100,7 +101,7 @@ SEXP eval_filter_one(SEXP quos, SEXP mask, SEXP caller, R_xlen_t n, SEXP env_fil
 
   // reduce
   R_xlen_t nquos = XLENGTH(quos);
-  for (R_xlen_t i=0; i < nquos; i++) {
+  for (R_xlen_t i = 0; i < nquos; i++) {
     SEXP current_expression = PROTECT(Rf_ScalarInteger(i+1));
     Rf_defineVar(dplyr::symbols::current_expression, current_expression, env_filter);
 
@@ -113,8 +114,9 @@ SEXP eval_filter_one(SEXP quos, SEXP mask, SEXP caller, R_xlen_t n, SEXP env_fil
       reduce_lgl(reduced, res, n);
     } else if(Rf_inherits(res, "data.frame")) {
       R_xlen_t ncol = XLENGTH(res);
-      for (R_xlen_t j=0; j<ncol; j++) {
-        reduce_lgl(reduced, VECTOR_ELT(res, j), n);
+      const SEXP* p_res = VECTOR_PTR_RO(res);
+      for (R_xlen_t j = 0; j<ncol; j++) {
+        reduce_lgl(reduced, p_res[j], n);
       }
     }
 
@@ -127,6 +129,7 @@ SEXP eval_filter_one(SEXP quos, SEXP mask, SEXP caller, R_xlen_t n, SEXP env_fil
 
 SEXP dplyr_mask_eval_all_filter(SEXP quos, SEXP env_private, SEXP s_n, SEXP env_filter) {
   DPLYR_MASK_INIT();
+  const SEXP* p_rows = VECTOR_PTR_RO(rows);
 
   R_xlen_t n = Rf_asInteger(s_n);
   SEXP keep = PROTECT(Rf_allocVector(LGLSXP, n));
@@ -134,7 +137,7 @@ SEXP dplyr_mask_eval_all_filter(SEXP quos, SEXP env_private, SEXP s_n, SEXP env_
 
   for (R_xlen_t i = 0; i < ngroups; i++) {
     DPLYR_MASK_SET_GROUP(i);
-    SEXP rows_i = VECTOR_ELT(rows, i);
+    SEXP rows_i = p_rows[i];
     R_xlen_t n_i = XLENGTH(rows_i);
 
     SEXP result_i = PROTECT(eval_filter_one(quos, mask, caller, n_i, env_filter));