+++ tests/testthat/test-group_data.R
@@ -62,10 +62,17 @@ test_that("no arg group_indices() is deprecated", {
 })
 
 test_that("group_indices(...) is deprecated", {
-  skip("Non-deterministic failures")
+  rlang::local_options(lifecycle_verbosity = "error")
 
   df <- tibble(x = 1, y = 2)
-  expect_warning(out <- df %>% group_indices(x), "deprecated")
+  expect_error(df %>% group_indices(x), "deprecated")
+})
+
+test_that("group_indices(...) still works though", {
+  rlang::local_options(lifecycle_verbosity = "quiet")
+
+  df <- tibble(x = 1, y = 2)
+  out <- df %>% group_indices(x)
   expect_equal(out, 1)
 })