+++ NEWS.md
@@ -1,11 +1,31 @@
 # dplyr (development version)
 
+* `inner_join()` gains a `keep` parameter for consistency with the other
+  mutating joins (@patrickbarks, #5581).
+
+* Improved performance with many columns, with a dynamic data mask using active
+  bindings and lazy chops (#5017). 
+
+* `mutate()` and friends preserves row names in data frames once more (#5418).
+
+* `group_by()` uses the ungrouped data for the implicit mutate step (#5598). 
+  You might have to define an `ungroup()` method for custom classes.
+  For example, see https://github.com/hadley/cubelyr/pull/3. 
+
+* `relocate()` can rename columns it relocates (#5569).
+
+* `distinct()` and `group_by()` have better error messages when the mutate step fails (#5060).
+
 * Clarify that `between()` is not vectorised (#5493).
 
 * Fixed `across()` issue where data frame columns would could not be referred to
   with `all_of()` in the nested case (`mutate()` within `mutate()`) (#5498).
   
 * `across()` handles data frames with 0 columns (#5523). 
+
+* `mutate()` always keeps grouping variables, unconditional to `.keep=` (#5582).
+
+* dplyr now depends on R 3.3.0
 
 # dplyr 1.0.2